refactor: extract options debug log export UI
